// raindrop-backend/src/vfs-api/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, Bucket, KvCache, Logger, MRNObject, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  logger: Logger;
  mem: KvCache;
  tracer: Tracer;
  VFS_API_KEY: string;
  VFS_STORAGE: Bucket;
}